# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Api
 * Secure E2E encrypted chat API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SignupInput {
  /**
     * @minLength 3
     * @maxLength 32
     */
  username: string;
  /** @minLength 6 */
  password: string;
  /** JWK-encoded RSA public key */
  publicKey: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface UserSummary {
  id: number;
  username: string;
  online: boolean;
  /** @nullable */
  lastSeen?: string | null;
  /** @nullable */
  profilePicture?: string | null;
}

export interface AuthResponse {
  token: string;
  user: UserSummary;
}

export interface PublicKeyResponse {
  username: string;
  publicKey: string;
}

export type MessageReactionsItem = {
  username: string;
  emoji: string;
};

export interface Message {
  id: number;
  fromUsername: string;
  toUsername: string;
  /** Base64-encoded encrypted ciphertext (never plaintext) */
  encryptedContent: string;
  /**
     * Base64-encoded AES key encrypted with recipient's RSA public key
     * @nullable
     */
  encryptedKey?: string | null;
  /**
     * Base64-encoded AES-GCM IV
     * @nullable
     */
  iv?: string | null;
  timestamp: string;
  delivered: boolean;
  read: boolean;
  reactions?: MessageReactionsItem[];
}

export interface ErrorResponse {
  error: string;
}