# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Secure E2E encrypted chat API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: users
    description: User operations
  - name: messages
    description: Message operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login to existing account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"

  /users/{username}/public-key:
    get:
      operationId: getPublicKey
      tags: [users]
      summary: Get a user's public key for encryption
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicKeyResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /messages/{username}:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get encrypted message history with a user
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

  /messages/unread:
    get:
      operationId: getUnreadMessages
      tags: [messages]
      summary: Get all unread (offline-stored) encrypted messages
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    SignupInput:
      type: object
      required: [username, password, publicKey]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
        password:
          type: string
          minLength: 6
        publicKey:
          type: string
          description: JWK-encoded RSA public key

    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserSummary"

    UserSummary:
      type: object
      required: [id, username, online]
      properties:
        id:
          type: integer
        username:
          type: string
        online:
          type: boolean
        lastSeen:
          type: ["string", "null"]
        profilePicture:
          type: ["string", "null"]

    PublicKeyResponse:
      type: object
      required: [username, publicKey]
      properties:
        username:
          type: string
        publicKey:
          type: string

    Message:
      type: object
      required: [id, fromUsername, toUsername, encryptedContent, timestamp, delivered, read]
      properties:
        id:
          type: integer
        fromUsername:
          type: string
        toUsername:
          type: string
        encryptedContent:
          type: string
          description: Base64-encoded encrypted ciphertext (never plaintext)
        encryptedKey:
          type: ["string", "null"]
          description: Base64-encoded AES key encrypted with recipient's RSA public key
        iv:
          type: ["string", "null"]
          description: Base64-encoded AES-GCM IV
        timestamp:
          type: string
        delivered:
          type: boolean
        read:
          type: boolean
        reactions:
          type: array
          items:
            type: object
            required: [username, emoji]
            properties:
              username:
                type: string
              emoji:
                type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string