# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Api
 * Secure E2E encrypted chat API
 * OpenAPI spec version: 0.1.0
 */
import type { MessageReactionsItem } from './messageReactionsItem';

export interface Message {
  id: number;
  fromUsername: string;
  toUsername: string;
  /** Base64-encoded encrypted ciphertext (never plaintext) */
  encryptedContent: string;
  /**
     * Base64-encoded AES key encrypted with recipient's RSA public key
     * @nullable
     */
  encryptedKey?: string | null;
  /**
     * Base64-encoded AES-GCM IV
     * @nullable
     */
  iv?: string | null;
  timestamp: string;
  delivered: boolean;
  read: boolean;
  reactions?: MessageReactionsItem[];
}