# lib/api-zod/src/generated/types/userSummary.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Api
 * Secure E2E encrypted chat API
 * OpenAPI spec version: 0.1.0
 */

export interface UserSummary {
  id: number;
  username: string;
  online: boolean;
  /** @nullable */
  lastSeen?: string | null;
  /** @nullable */
  profilePicture?: string | null;
}